fixed the fourth task, second homework is done

+++ 2.Homeworks/2.Second/src/HW_2_4/FourthTask.java
@@ -12,7 +12,7 @@ public class FourthTask {
         StringBuilder answer = new StringBuilder();
         int q;
         int n;
-        int currentInversions;
+        long currentInversions;
         List<Integer> integerList = new ArrayList<>();
         InversionsCounter<Integer> inversionsCounter = new InversionsCounter<>();
